app/requirements.txt
streamlit>=1.30.0
openai>=1.0.0
PyMuPDF>=1.23.0
easyocr>=1.7.0
jinja2>=3.1.0
playwright>=1.40.0
markdown>=3.5.0
python-dotenv>=1.0.0
